Test adjustments and client implementation
