+++ src/Explorer/QueryCopilot/QueryCopilotUtilities.ts
@@ -1,66 +1,11 @@
 import { FeedOptions, Item, ItemDefinition, QueryIterator, Resource } from "@azure/cosmos";
-import {
-  QueryCopilotSampleContainerId,
-  QueryCopilotSampleContainerSchema,
-  QueryCopilotSampleDatabaseId,
-  ShortenedQueryCopilotSampleContainerSchema,
-} from "Common/Constants";
+import { QueryCopilotSampleContainerId, QueryCopilotSampleDatabaseId } from "Common/Constants";
 import { handleError } from "Common/ErrorHandlingUtils";
 import { sampleDataClient } from "Common/SampleDataClient";
-import { createUri } from "Common/UrlUtility";
 import { getPartitionKeyValue } from "Common/dataAccess/getPartitionKeyValue";
 import { getCommonQueryOptions } from "Common/dataAccess/queryDocuments";
-import Explorer from "Explorer/Explorer";
-import { useNotebook } from "Explorer/Notebook/useNotebook";
-import { FeedbackParams } from "Explorer/QueryCopilot/Shared/QueryCopilotInterfaces";
 import DocumentId from "Explorer/Tree/DocumentId";
-import { userContext } from "UserContext";
 import { logConsoleProgress } from "Utils/NotificationConsoleUtils";
-import { useQueryCopilot } from "hooks/useQueryCopilot";
-
-export const submitFeedback = async ({
-  params,
-  explorer,
-}: {
-  params: FeedbackParams;
-  explorer: Explorer;
-}): Promise<void> => {
-  try {
-    const { likeQuery, generatedQuery, userPrompt, description, contact } = params;
-    const { correlationId, shouldAllocateContainer, setShouldAllocateContainer } = useQueryCopilot();
-    const payload = {
-      containerSchema: userContext.features.enableCopilotFullSchema
-        ? QueryCopilotSampleContainerSchema
-        : ShortenedQueryCopilotSampleContainerSchema,
-      like: likeQuery ? "like" : "dislike",
-      generatedSql: generatedQuery,
-      userPrompt,
-      description: description || "",
-      contact: contact || "",
-    };
-    if (shouldAllocateContainer && userContext.features.enableCopilotPhoenixGateaway) {
-      await explorer.allocateContainer();
-      setShouldAllocateContainer(false);
-    }
-    const serverInfo = useNotebook.getState().notebookServerInfo;
-    const feedbackUri = userContext.features.enableCopilotPhoenixGateaway
-      ? createUri(serverInfo.notebookServerEndpoint, "feedback")
-      : createUri("https://copilotorchestrater.azurewebsites.net/", "feedback");
-    const response = await fetch(feedbackUri, {
-      method: "POST",
-      headers: {
-        "content-type": "application/json",
-        "x-ms-correlationid": correlationId,
-      },
-      body: JSON.stringify(payload),
-    });
-    if (response.status === 404) {
-      setShouldAllocateContainer(true);
-    }
-  } catch (error) {
-    handleError(error, "copilotSubmitFeedback");
-  }
-};
 
 export const querySampleDocuments = (query: string, options: FeedOptions): QueryIterator<ItemDefinition & Resource> => {
   options = getCommonQueryOptions(options);
